pyproject.toml
[project]
name = "live-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioconsole>=0.8.2",
    "google-genai>=1.56.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.0",
]
